--- src/main/java/com/DrWait/core/security/auth/dto/LoginRequestDto.java
@@ -1,10 +1,0 @@
-package com.DrWait.core.security.auth.dto;
-
-import lombok.Getter;
-
-@Getter
-public class LoginRequestDto {
-
-    private String username;
-    private String password;
-}
