--- src/main/java/com/DrWait/core/security/auth/dto/SignupRequestDto.java
@@ -1,12 +1,0 @@
-package com.DrWait.core.security.auth.dto;
-
-import lombok.Getter;
-
-@Getter
-public class SignupRequestDto {
-
-    private String username;
-    private String password;
-    private String phoneNumber;
-    private String residentRegistrationNumber;
-}
