--- src/main/java/com/DrWait/core/security/token/JwtAccessDeniedHandler.java
@@ -1,26 +1,0 @@
-package com.DrWait.core.security.token;
-
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import org.springframework.security.access.AccessDeniedException;
-import org.springframework.security.web.access.AccessDeniedHandler;
-import org.springframework.stereotype.Component;
-
-import java.io.IOException;
-
-/* 🔒 인가 실패 핸들러 */
-@Component
-public class JwtAccessDeniedHandler implements AccessDeniedHandler {
-
-    @Override
-    public void handle(
-            HttpServletRequest request,
-            HttpServletResponse response,
-            AccessDeniedException accessDeniedException
-    ) throws IOException {
-        response.sendError(HttpServletResponse.SC_FORBIDDEN, "접근 권한이 없습니다.");
-    }
-
-    public static class JwtConfig {
-    }
-}
