--- src/main/java/com/DrWait/core/security/token/JwtAuthenticationFilter.java
@@ -1,52 +1,0 @@
-package com.DrWait.core.security.token;
-
-import com.DrWait.domain.user.service.CustomUserDetailsService;
-import jakarta.servlet.FilterChain;
-import jakarta.servlet.ServletException;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.web.authentication.WebAuthenticationDetailsSource;
-import org.springframework.stereotype.Component;
-import org.springframework.web.filter.OncePerRequestFilter;
-
-import java.io.IOException;
-
-/* 🔒 JWT 인증 필터 */
-/// 매 요청마다 JWT 토큰을 검사
-/// 검증된 사용자 정보를 Spring Security의 SecurityContext에 등록해주는 역활
-@Slf4j
-@RequiredArgsConstructor
-@Component
-public class JwtAuthenticationFilter extends OncePerRequestFilter {
-
-    private final JwtTokenProvider jwtTokenProvider;
-    private final CustomUserDetailsService userDetailsService;
-
-    @Override
-    protected void doFilterInternal(
-            HttpServletRequest request,
-            HttpServletResponse response,
-            FilterChain filterChain) throws ServletException, IOException {
-
-        // 토큰 추출
-        String token = jwtTokenProvider.resolveToken(request);
-
-        // 토큰 유효성 검사
-        if (token != null && jwtTokenProvider.validateToken(token)) {
-            var authentication = jwtTokenProvider.getAuthentication(token);
-
-            if(authentication instanceof UsernamePasswordAuthenticationToken authToken){
-                authToken.setDetails(new WebAuthenticationDetailsSource().buildDetails(request));
-                SecurityContextHolder.getContext().setAuthentication(authentication);
-
-                log.info("[JwtAuthenticationFilter] 인증 성공: {}", authentication.getName());
-            }
-        }
-
-        filterChain.doFilter(request, response);
-    }
-}
