+++ src/main/java/com/DrWait/domain/family/controller/FamilyController.java
@@ -1,0 +1,11 @@
+package com.DrWait.domain.family.controller;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequiredArgsConstructor
+@RequestMapping("api/family")
+public class FamilyController {
+}
