+++ src/main/java/com/DrWait/domain/family/entity/FamilyGroup.java
@@ -1,0 +1,21 @@
+package com.DrWait.domain.family.entity;
+
+import jakarta.persistence.*;
+import lombok.*;
+
+import java.util.UUID;
+
+@Entity
+@Getter @Setter
+@Builder
+@AllArgsConstructor
+@RequiredArgsConstructor(access = AccessLevel.PROTECTED)
+@Table(name = "family_group")
+public class FamilyGroup {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    private UUID userId;
+}
