+++ src/main/java/com/DrWait/domain/family/entity/FamilyMemberId.java
@@ -1,0 +1,17 @@
+package com.DrWait.domain.family.entity;
+
+import jakarta.persistence.Embeddable;
+import lombok.*;
+
+import java.util.UUID;
+
+@Embeddable
+@Getter @Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@EqualsAndHashCode
+public class FamilyMemberId {
+
+    private UUID userId;
+    private Long familyGroupId;
+}
