+++ src/main/java/com/DrWait/domain/family/repository/FamilyGroupRepository.java
@@ -1,0 +1,13 @@
+package com.DrWait.domain.family.repository;
+
+import com.DrWait.domain.family.entity.FamilyGroup;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.Optional;
+
+@Repository
+public interface FamilyGroupRepository extends JpaRepository<FamilyGroup, Long> {
+
+    Optional<FamilyGroup> findById(Long id);
+}
