--- src/main/java/com/DrWait/domain/user/enums/Role.java
@@ -1,5 +1,0 @@
-package com.DrWait.domain.user.enums;
-
-public enum Role {
-    USER, ADMIN
-}
