+++ src/main/java/com/DrWait/domain/user/service/UserService.java
@@ -3,6 +3,8 @@ package com.DrWait.domain.user.service;
 import com.DrWait.domain.user.dto.UserInfoResponseDto;
 import com.DrWait.domain.user.entity.User;
 import com.DrWait.domain.user.repository.UserRepository;
+import com.DrWait.global.error.CustomException;
+import com.DrWait.global.error.ErrorCode;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
@@ -17,12 +19,17 @@ public class UserService {
 
     private final UserRepository userRepository;
 
-    public UserInfoResponseDto getUserInfo(String userId){
-        log.info("서치 시작");
+    public User getUserEntityByUserId(String userId){
+        return userRepository.findById(UUID.fromString(userId))
+                .orElseThrow(() -> new CustomException(ErrorCode.USER_NOT_FOUND));
+    }
 
-        User user = userRepository.findById(UUID.fromString(userId))
-                .orElseThrow(() -> new UsernameNotFoundException("사용자를 찾을 수 없습니다."));
+    public User getUserEntityByUsername(String username){
+        return userRepository.findByUsername(username)
+                .orElseThrow(() -> new CustomException(ErrorCode.USER_NOT_FOUND));
+    }
 
-        return new UserInfoResponseDto(user);
+    public UserInfoResponseDto getUserByUserId(String userId){
+        return new UserInfoResponseDto(getUserEntityByUserId(userId));
     }
 }
