+++ src/main/java/com/DrWait/global/error/CustomException.java
@@ -1,0 +1,10 @@
+package com.DrWait.global.error;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class CustomException extends RuntimeException {
+    ErrorCode errorCode;
+}
