+++ src/main/java/com/DrWait/global/security/auth/dto/HospitalSignupRequestDto.java
@@ -1,0 +1,15 @@
+package com.DrWait.global.security.auth.dto;
+
+import lombok.Getter;
+
+@Getter
+public class HospitalSignupRequestDto {
+
+    private String name;
+    private String department;
+    private String address;
+    private String telephone;
+    private String website_url;
+    private String username;
+    private String password;
+}
