+++ src/main/java/com/DrWait/global/security/auth/dto/LoginRequestDto.java
@@ -1,0 +1,10 @@
+package com.DrWait.global.security.auth.dto;
+
+import lombok.Getter;
+
+@Getter
+public class LoginRequestDto {
+
+    private String username;
+    private String password;
+}
