+++ src/main/java/com/DrWait/global/security/auth/enums/Role.java
@@ -1,0 +1,5 @@
+package com.DrWait.global.security.auth.enums;
+
+public enum Role {
+    USER, ADMIN
+}
